Corrected compile issues on BalanceFinder.
When solving the compile issues, the balance hash calculation was
changed in the code. The tool cannnot create the same balance
hashes as factomd does without more work to emulate the factoid state.

If we use the factoid state in the tool, that would ruin the point of
the tool checking the balances. As this tool calculates the balances
with different logic.