Corrected processing of a message after going into the processlist
If a leader adds a message to it's processlist, it will immediately
process it such that the next message can be executed, and not held
in holding.

Also added code to dump prometheus counters for testing purposes.
In TestLoad() this code reduces the amount of things added to holding
by the total number of transactions in the test (~2k)